regen(go-src-converted): rebank the full stdlib with -tags purego + converter fixes
Regenerates the committed full-conversion tree from the current converter,
which now applies -tags purego by default for -stdlib. This makes the
published corpus reproduce Go built with -tags purego, as the README claims:
the portable pure-Go crypto/hash variants (which a source transpiler can
convert) replace the bodyless assembly declarations that compiled to
throwing stubs.

Net change: 17 asm-variant .cs removed, 15 portable variants added (e.g.
sha256block_amd64.cs -> sha256block_generic.cs, poly1305/sum_amd64.cs ->
mac_noasm.cs, sha3/keccakf_amd64.cs -> keccakf.cs), and ~700 production .cs
updated to capture the converter fixes landed since the 2026-07-17 rebank
(Go-semantics shift, else-if restoration, interface-assertion records,
fixed-array construction, cross-package field symbol resolution, the
select-send readiness guard, and the builtin-shadow / range-var fixes).
crypto/sha256 now computes a real SHA-256 instead of throwing.

Regenerable and byte-deterministic: two full `-comments -stdlib` reconverts
diff to zero. Hand-owned files preserved (21 *_impl.cs, 20 whole-file
GoManualConversion .cs, the unsafe/ package, the 15 validated packages'
committed test artifacts + csproj exclusions, and the 3 disclosure
manifests); the crypto csproj dependency sets follow the purego variants
(subtle/maphash/bigmod/sha3/alias gain refs, the asm packages shed
internal/cpu and friends). Full corpus builds 302/302 with zero errors.

+++ src/go-src-converted/bufio/scan.cs
@@ -167,7 +167,7 @@ public static error ErrFinalToken = errors.New("final token"u8);
         if (s.end == len(s.buf)) {
             // Guarantee no overflow in the multiplication below.
             nint maxInt = /* int(^uint(0) >> 1) */ unchecked((nint)9223372036854775807);
-            if (len(s.buf) >= s.maxTokenSize || len(s.buf) > 4611686018427387903L) {
+            if (len(s.buf) >= s.maxTokenSize || len(s.buf) > (nint)(4611686018427387903L)) {
                 s.setErr(ErrTooLong);
                 return false;
             }
