+++ src/go-src-converted/bytes/bytes.cs
@@ -1209,7 +1209,7 @@ public static bool EqualFold(slice<byte> s, slice<byte> t) {
             (tr, sr) = (sr, tr);
         }
         // ASCII only, sr/tr must be upper/lower case
-        if ((rune)'A' <= sr && sr <= (rune)'Z' && tr == sr + (rune)'a' - (rune)'A') {
+        if ((rune)'A' <= sr && sr <= (rune)'Z' && tr == (byte)(sr + (rune)'a' - (rune)'A')) {
             continue;
         }
         return false;
