+++ src/go-src-converted/unicode/letter.cs
@@ -94,7 +94,7 @@ internal static bool is16(slice<Range16> ranges, uint16 r) {
                 return false;
             }
             if (r <= (~range_).Hi) {
-                return (~range_).Stride == 1 || (r - (~range_).Lo) % (~range_).Stride == 0;
+                return (~range_).Stride == 1 || (uint16)((r - (~range_).Lo) % (~range_).Stride) == 0;
             }
         }
         return false;
@@ -106,7 +106,7 @@ internal static bool is16(slice<Range16> ranges, uint16 r) {
         nint m = (nint)(((nuint)(lo + hi) >> (int)(1)));
         var range_ = Ꮡ(ranges, m);
         if ((~range_).Lo <= r && r <= (~range_).Hi) {
-            return (~range_).Stride == 1 || (r - (~range_).Lo) % (~range_).Stride == 0;
+            return (~range_).Stride == 1 || (uint16)((r - (~range_).Lo) % (~range_).Stride) == 0;
         }
         if (r < (~range_).Lo){
             hi = m;
